Cleanup high level graph tests
