+++ grakn-graph/src/test/java/ai/grakn/graph/internal/RelationTest.java
@@ -18,6 +18,7 @@
 
 package ai.grakn.graph.internal;
 
+import ai.grakn.concept.Concept;
 import ai.grakn.concept.Entity;
 import ai.grakn.concept.EntityType;
 import ai.grakn.concept.Instance;
@@ -26,10 +27,12 @@ import ai.grakn.concept.RelationType;
 import ai.grakn.concept.Resource;
 import ai.grakn.concept.ResourceType;
 import ai.grakn.concept.RoleType;
+import ai.grakn.concept.Type;
 import ai.grakn.exception.ConceptException;
 import ai.grakn.exception.GraknValidationException;
 import ai.grakn.util.ErrorMessage;
 import ai.grakn.util.Schema;
+import com.google.common.collect.Iterators;
 import org.apache.tinkerpop.gremlin.structure.Direction;
 import org.apache.tinkerpop.gremlin.structure.Edge;
 import org.apache.tinkerpop.gremlin.structure.Vertex;
@@ -38,15 +41,26 @@ import org.junit.Test;
 
 import java.util.Collection;
 import java.util.HashMap;
+import java.util.HashSet;
 import java.util.Iterator;
+import java.util.List;
 import java.util.Map;
 import java.util.Set;
 import java.util.TreeMap;
+import java.util.stream.Collectors;
 
+import static ai.grakn.util.Schema.EdgeProperty.FROM_ID;
+import static ai.grakn.util.Schema.EdgeProperty.FROM_TYPE_NAME;
+import static ai.grakn.util.Schema.EdgeProperty.RELATION_ID;
+import static ai.grakn.util.Schema.EdgeProperty.RELATION_TYPE_NAME;
+import static ai.grakn.util.Schema.EdgeProperty.TO_ID;
+import static ai.grakn.util.Schema.EdgeProperty.TO_ROLE_NAME;
+import static ai.grakn.util.Schema.EdgeProperty.TO_TYPE_NAME;
 import static org.hamcrest.CoreMatchers.allOf;
 import static org.hamcrest.CoreMatchers.containsString;
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertFalse;
+import static org.junit.Assert.assertNotNull;
 import static org.junit.Assert.assertNull;
 import static org.junit.Assert.assertTrue;
 
@@ -96,7 +110,50 @@ public class RelationTest extends GraphTestBase{
     }
 
     @Test
-    public void scopeTest(){
+    public void testShortcutEdges(){
+        EntityImpl a = (EntityImpl) type.addEntity();
+        EntityImpl b = (EntityImpl) type.addEntity();
+        EntityImpl c = (EntityImpl) type.addEntity();
+
+        assertEquals(0, a.getOutgoingNeighbours(Schema.EdgeLabel.SHORTCUT).size());
+        assertEquals(0, b.getOutgoingNeighbours(Schema.EdgeLabel.SHORTCUT).size());
+        assertEquals(0, c.getOutgoingNeighbours(Schema.EdgeLabel.SHORTCUT).size());
+
+        Relation relation = relationType.addRelation();
+
+        relation.putRolePlayer(role1, a);
+        assertEquals(0, a.getOutgoingNeighbours(Schema.EdgeLabel.SHORTCUT).size());
+        assertEquals(0, b.getOutgoingNeighbours(Schema.EdgeLabel.SHORTCUT).size());
+        assertEquals(0, c.getOutgoingNeighbours(Schema.EdgeLabel.SHORTCUT).size());
+
+        relation.putRolePlayer(role2, b);
+        assertEquals(1, a.getOutgoingNeighbours(Schema.EdgeLabel.SHORTCUT).size());
+        assertEquals(1, b.getOutgoingNeighbours(Schema.EdgeLabel.SHORTCUT).size());
+        assertEquals(0, c.getOutgoingNeighbours(Schema.EdgeLabel.SHORTCUT).size());
+
+        relation.putRolePlayer(role3, c);
+        assertEquals(2, a.getOutgoingNeighbours(Schema.EdgeLabel.SHORTCUT).size());
+        assertEquals(2, b.getOutgoingNeighbours(Schema.EdgeLabel.SHORTCUT).size());
+        assertEquals(2, c.getOutgoingNeighbours(Schema.EdgeLabel.SHORTCUT).size());
+
+        //Check Structure of Shortcut Edge
+        boolean foundEdge = false;
+        for (EdgeImpl edge : a.getEdgesOfType(Direction.OUT, Schema.EdgeLabel.SHORTCUT)) {
+            if(edge.getProperty(TO_ROLE_NAME).equals(role2.getName())) {
+                foundEdge = true;
+                assertEquals(edge.getProperty(RELATION_TYPE_NAME), relationType.getName());
+                assertEquals(edge.getProperty(RELATION_ID), relation.getId());
+                assertEquals(edge.getProperty(TO_ID), b.getId());
+                assertEquals(edge.getProperty(FROM_ID), a.getId());
+                assertEquals(edge.getProperty(FROM_TYPE_NAME), a.type().getName());
+                assertEquals(edge.getProperty(TO_TYPE_NAME), b.type().getName());
+            }
+        }
+        assertTrue(foundEdge);
+    }
+
+    @Test
+    public void testScope(){
         RelationType relationType = graknGraph.putRelationType("rel type");
         RelationImpl relation = (RelationImpl) relationType.addRelation();
         RelationImpl relationValue = (RelationImpl) relationType.addRelation();
@@ -211,7 +268,7 @@ public class RelationTest extends GraphTestBase{
         Relation rel = relationType.addRelation().
                 putRolePlayer(roleType1, instance1).putRolePlayer(roleType2, instance2).putRolePlayer(roleType3, resource);
 
-        Relation rel2 = relationType.addRelation().putRolePlayer(roleType1, instance1).putRolePlayer(roleType2, instance3);
+        relationType.addRelation().putRolePlayer(roleType1, instance1).putRolePlayer(roleType2, instance3);
 
         assertEquals(1, instance1.resources().size());
         assertEquals(2, resource.ownerInstances().size());
@@ -239,7 +296,6 @@ public class RelationTest extends GraphTestBase{
         graknGraph.commit();
 
         relation = (RelationImpl) graknGraph.getRelationType("relation type").instances().iterator().next();
-        //assertTrue(relation.getIndex().startsWith("RelationBaseId_" + relation.getBaseIdentifier()));
 
         roleType1 = graknGraph.putRoleType("role type 1");
         roleType2 = graknGraph.putRoleType("role type 2");
@@ -276,7 +332,7 @@ public class RelationTest extends GraphTestBase{
         Instance instance2 = type.addEntity();
 
         Relation relation1 = relationType.addRelation().putRolePlayer(roleType1, instance1).putRolePlayer(roleType2, instance2);
-        Relation relation2 = relationType.addRelation().putRolePlayer(roleType1, instance1).putRolePlayer(roleType2, instance2);
+        relationType.addRelation().putRolePlayer(roleType1, instance1).putRolePlayer(roleType2, instance2);
 
         expectedException.expect(GraknValidationException.class);
         expectedException.expectMessage(allOf(
@@ -304,7 +360,7 @@ public class RelationTest extends GraphTestBase{
     }
 
     @Test
-    public void makeSureCastingsNotRemoved(){
+    public void testMakeSureCastingsNotRemoved(){
         RoleType entityRole = graknGraph.putRoleType("Entity Role");
         RoleType degreeRole = graknGraph.putRoleType("Degree Role");
         EntityType entityType = graknGraph.putEntityType("Entity Type").playsRole(entityRole);
@@ -345,4 +401,203 @@ public class RelationTest extends GraphTestBase{
         assertTrue("Relation toString missing role and role player definition", relation.toString().contains(rolerp1));
         assertTrue("Relation toString missing role and role player definition", relation.toString().contains(rolerp2));
     }
+
+    @Test
+    public void testDeleteOfEachRolePlayerAndAutoCleanupOfRelation(){
+        RoleType roleA = graknGraph.putRoleType("RoleA");
+        RoleType roleB = graknGraph.putRoleType("RoleB");
+        RoleType roleC = graknGraph.putRoleType("RoleC");
+
+        RelationType relation = graknGraph.putRelationType("relation type").hasRole(roleA).hasRole(roleB).hasRole(roleC);
+        EntityType type = graknGraph.putEntityType("concept type").playsRole(roleA).playsRole(roleB).playsRole(roleC);
+        Entity a = type.addEntity();
+        Entity b = type.addEntity();
+        Entity c = type.addEntity();
+
+        String relationId = relation.addRelation().putRolePlayer(roleA, a).putRolePlayer(roleB, b).putRolePlayer(roleC, c).getId();
+
+        a.delete();
+        assertNotNull(graknGraph.getConcept(relationId));
+        b.delete();
+        assertNotNull(graknGraph.getConcept(relationId));
+        c.delete();
+        assertNull(graknGraph.getConcept(relationId));
+    }
+
+    @Test
+    public void testAddRolePlayerToExistingRelation(){
+        RelationType cast = graknGraph.putRelationType("Cast");
+        RoleType feature = graknGraph.putRoleType("Feature");
+        RoleType actor = graknGraph.putRoleType("Actor");
+        EntityType movie = graknGraph.putEntityType("Movie");
+        EntityType person = graknGraph.putEntityType("Person");
+        Instance pacino = person.addEntity();
+        Instance godfather = movie.addEntity();
+
+        RelationImpl relation = (RelationImpl) cast.addRelation().
+                putRolePlayer(feature, null).putRolePlayer(actor, pacino);
+        relation.putRolePlayer(feature, godfather);
+        assertEquals(2, relation.getMappingCasting().size());
+        assertTrue(relation.rolePlayers().values().contains(pacino));
+        assertTrue(relation.rolePlayers().values().contains(godfather));
+    }
+
+    @Test
+    public void testGetRelationWithMissingRolePlayers(){
+        RelationType relationType2 = graknGraph.putRelationType("relationType2");
+
+        Map<RoleType, Instance> roleMap = new HashMap<>();
+        roleMap.put(role1, null);
+        roleMap.put(role2, rolePlayer2);
+        roleMap.put(role3, null);
+
+        relationType.hasRole(role1);
+        relationType.hasRole(role2);
+        relationType.hasRole(role3);
+        relationType.addRelation().
+                putRolePlayer(role1, null).putRolePlayer(role2, rolePlayer2).putRolePlayer(role3, null);
+
+        assertNotNull(graknGraph.getRelation(relationType, roleMap));
+        assertNull(graknGraph.getRelation(relationType2, roleMap));
+    }
+
+    @Test
+    public void testAddRelationshipWithNullRole(){
+        expectedException.expect(RuntimeException.class);
+        expectedException.expectMessage(allOf(
+                containsString(ErrorMessage.ROLE_IS_NULL.getMessage(rolePlayer1))
+        ));
+
+        relationType.addRelation().putRolePlayer(null, rolePlayer1);
+    }
+
+    @Test
+    public void testAddLargeAndMultipleRelationships(){
+
+        //Actual Concepts To Appear Linked In Graph
+        Type relationType = graknGraph.putEntityType("Relation Type");
+        RelationTypeImpl cast = (RelationTypeImpl) graknGraph.putRelationType("Cast");
+        EntityType roleType = graknGraph.putEntityType("Role Type");
+        EntityType type = graknGraph.putEntityType("Concept Type");
+        RoleTypeImpl feature = (RoleTypeImpl) graknGraph.putRoleType("Feature");
+        RoleTypeImpl actor = (RoleTypeImpl) graknGraph.putRoleType("Actor");
+        EntityType movie = graknGraph.putEntityType("Movie");
+        EntityType person = graknGraph.putEntityType("Person");
+        InstanceImpl pacino = (InstanceImpl) person.addEntity();
+        InstanceImpl godfather = (InstanceImpl) movie.addEntity();
+        EntityType genre = graknGraph.putEntityType("Genre");
+        RoleTypeImpl movieOfGenre = (RoleTypeImpl) graknGraph.putRoleType("Movie of Genre");
+        RoleTypeImpl movieGenre = (RoleTypeImpl) graknGraph.putRoleType("Movie Genre");
+        InstanceImpl crime = (InstanceImpl) genre.addEntity();
+        RelationTypeImpl movieHasGenre = (RelationTypeImpl) graknGraph.putRelationType("Movie Has Genre");
+
+        //Construction
+        cast.hasRole(feature);
+        cast.hasRole(actor);
+
+        RelationImpl assertion = (RelationImpl) cast.addRelation().
+                putRolePlayer(feature, godfather).putRolePlayer(actor, pacino);
+
+        Map<RoleType, Instance> roleMap = new HashMap<>();
+        roleMap.clear();
+        roleMap.put(movieOfGenre, godfather);
+        roleMap.put(movieGenre, crime);
+        movieHasGenre.addRelation().
+                putRolePlayer(movieOfGenre, godfather).putRolePlayer(movieGenre, crime);
+
+        movieHasGenre.hasRole(movieOfGenre);
+        movieHasGenre.hasRole(movieGenre);
+
+        //Validation
+        //Counts
+        assertEdgeCountOfVertex(type, Schema.EdgeLabel.SUB, 0, 1);
+        assertEdgeCountOfVertex(relationType, Schema.EdgeLabel.SUB, 0, 1);
+        assertEdgeCountOfVertex(roleType, Schema.EdgeLabel.SUB, 0, 1);
+        assertEdgeCountOfVertex(cast, Schema.EdgeLabel.SUB, 0, 1);
+        assertEdgeCountOfVertex(cast, Schema.EdgeLabel.HAS_ROLE, 0, 2);
+
+        assertEdgeCountOfVertex(feature, Schema.EdgeLabel.SUB, 0, 1);
+        assertEdgeCountOfVertex(feature, Schema.EdgeLabel.CASTING, 0, 0);
+
+        assertEdgeCountOfVertex(actor, Schema.EdgeLabel.SUB, 0, 1);
+        assertEdgeCountOfVertex(actor, Schema.EdgeLabel.CASTING, 0, 0);
+
+        assertEdgeCountOfVertex(person, Schema.EdgeLabel.SUB, 0, 1);
+        assertEdgeCountOfVertex(movie, Schema.EdgeLabel.SUB, 0, 1);
+        assertEdgeCountOfVertex(crime, Schema.EdgeLabel.ISA, 0, 1);
+        assertEdgeCountOfVertex(genre, Schema.EdgeLabel.SUB, 0, 1);
+
+        assertEdgeCountOfVertex(pacino, Schema.EdgeLabel.ISA, 0, 1);
+        assertEdgeCountOfVertex(pacino, Schema.EdgeLabel.ROLE_PLAYER, 1, 0);
+
+        assertEdgeCountOfVertex(movieOfGenre, Schema.EdgeLabel.SUB, 0, 1);
+        assertEdgeCountOfVertex(movieOfGenre, Schema.EdgeLabel.ROLE_PLAYER, 0, 0);
+
+        assertEdgeCountOfVertex(movieGenre, Schema.EdgeLabel.SUB, 0, 1);
+        assertEdgeCountOfVertex(movieGenre, Schema.EdgeLabel.ROLE_PLAYER, 0, 0);
+
+        assertEdgeCountOfVertex(movieHasGenre, Schema.EdgeLabel.SUB, 0, 1);
+        assertEdgeCountOfVertex(movieHasGenre, Schema.EdgeLabel.HAS_ROLE, 0, 2);
+
+        assertEdgeCountOfVertex(godfather, Schema.EdgeLabel.ISA, 0, 1);
+        assertEdgeCountOfVertex(godfather, Schema.EdgeLabel.ROLE_PLAYER, 2, 0);
+
+        //More Specific Checks
+        List<Vertex> assertionsTypes = graknGraph.getTinkerPopGraph().traversal().V(godfather.getBaseIdentifier()).in(Schema.EdgeLabel.ROLE_PLAYER.getLabel()).in(Schema.EdgeLabel.CASTING.getLabel()).out(Schema.EdgeLabel.ISA.getLabel()).toList();
+        assertEquals(2, assertionsTypes.size());
+
+        List<Object> assertTypeIds = assertionsTypes.stream().map(Vertex::id).collect(Collectors.toList());
+
+        assertTrue(assertTypeIds.contains(cast.getBaseIdentifier()));
+        assertTrue(assertTypeIds.contains(movieHasGenre.getBaseIdentifier()));
+
+        List<Vertex> collection = graknGraph.getTinkerPopGraph().traversal().V(cast.getBaseIdentifier(), movieHasGenre.getBaseIdentifier()).in(Schema.EdgeLabel.ISA.getLabel()).out(Schema.EdgeLabel.CASTING.getLabel()).out().toList();
+        assertEquals(8, collection.size());
+
+        HashSet<Object> uniqueCollection = new HashSet<>();
+        for(Vertex v: collection){
+            uniqueCollection.add(v.id());
+        }
+
+        assertEquals(7, uniqueCollection.size());
+        assertTrue(uniqueCollection.contains(feature.getBaseIdentifier()));
+        assertTrue(uniqueCollection.contains(actor.getBaseIdentifier()));
+        assertTrue(uniqueCollection.contains(godfather.getBaseIdentifier()));
+        assertTrue(uniqueCollection.contains(pacino.getBaseIdentifier()));
+        assertTrue(uniqueCollection.contains(movieOfGenre.getBaseIdentifier()));
+        assertTrue(uniqueCollection.contains(movieGenre.getBaseIdentifier()));
+        assertTrue(uniqueCollection.contains(crime.getBaseIdentifier()));
+
+        assertEquals(Schema.BaseType.ENTITY.name(), pacino.getBaseType());
+        for(CastingImpl casting: assertion.getMappingCasting()){
+            assertEquals(casting.getRolePlayer().getBaseType(), Schema.BaseType.ENTITY.name());
+        }
+
+    }
+    private void assertEdgeCountOfVertex(Concept concept , Schema.EdgeLabel type, int inCount, int outCount){
+        Vertex v= graknGraph.getTinkerPopGraph().traversal().V(((ConceptImpl) concept).getBaseIdentifier()).next();
+        assertEquals(inCount, Iterators.size(v.edges(Direction.IN, type.getLabel())));
+        assertEquals(outCount, Iterators.size(v.edges(Direction.OUT, type.getLabel())));
+    }
+
+    @Test
+    public void testGetRelationViaMap() {
+        RoleType role1 = graknGraph.putRoleType("Role 1");
+        RoleType role2 = graknGraph.putRoleType("Role 2");
+        EntityType entityType = graknGraph.putEntityType("entityType").playsRole(role1).playsRole(role2);
+        RelationType relationType = graknGraph.putRelationType("relationType").hasRole(role1).hasRole(role2);
+
+        Entity instance1 = entityType.addEntity();
+        Entity instance2 = entityType.addEntity();
+        Relation relation = relationType.addRelation().
+                putRolePlayer(role1, instance1).
+                putRolePlayer(role2, instance2);
+
+        Map<RoleType, Instance> map = new HashMap<>();
+        map.put(role1, instance1);
+        map.put(role2, instance2);
+
+        assertEquals(relation, graknGraph.getRelation(relationType, map));
+    }
+
 }
